refactor(AnswerTile): Move into presenter folder
